pyproject.toml
[project]
name = "xylem-scm-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=3.0.2",
    "playwright>=1.58.0",
    "python-dotenv>=1.0.0",
]
